Experiment's Kubernetes cluster launchers moved to pkg/experiment
