+++ experiments/memcached-sensitivity-profile/main.go
@@ -62,8 +62,8 @@ func main() {
 	validate.OS()
 
 	// Launch Kubernetes cluster.
-	if sensitivity.ShouldLaunchKubernetesCluster() {
-		handle, err := sensitivity.LaunchKubernetesCluster()
+	if experiment.ShouldLaunchKubernetesCluster() {
+		handle, err := experiment.LaunchKubernetesCluster()
 		errutil.CheckWithContext(err, "Could not launch Kubernetes cluster")
 		defer handle.Stop()
 	}
