+++ experiments/optimal-core-allocation/main.go
@@ -104,8 +104,8 @@ func main() {
 	}
 
 	// Launch Kubernetes cluster.
-	if sensitivity.ShouldLaunchKubernetesCluster() {
-		handle, err := sensitivity.LaunchKubernetesCluster()
+	if experiment.ShouldLaunchKubernetesCluster() {
+		handle, err := experiment.LaunchKubernetesCluster()
 		errutil.CheckWithContext(err, "Could not launch Kubernetes cluster")
 		defer handle.Stop()
 	}
